backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Data / ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
shap==0.46.0
joblib==1.4.2

# AI / LangChain
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1

# Database
supabase==2.7.4

# HTTP & Utils
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# File handling
openpyxl==3.1.5
